sched: allow to run actions on the same goroutine
go-sched automatically run event actions on separate goroutine.
This don't play nicely with some usage scenarios, most notably fake
time.
We believe the policy which decides where the action should be run
should not be hardcoded in the scheduler, hence we provide options
to run the action code in the same scheduler goroutine.

Signed-off-by: Francesco Romani <fromani@redhat.com>